refactor job/task status logger; rename status_num
don't try and pretend that status_num means something. It's just a
arbitrary field to allow us to differentiate different kinds of
threshold based alerts we want to write. Just call it value and add
comments that explain what numbers mean what.
